% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meansTestPPair.R
\name{meansTestPPair}
\alias{meansTestPPair}
\title{Significance test for difference between 2 mean values of pair groups}
\usage{
meansTestPPair(nd1, av1, sd1, av2, sd2, sd3)
}
\arguments{
\item{nd1}{number of group}

\item{av1}{mean value of group 1}

\item{sd1}{standard deviation of group 1}

\item{av2}{mean value of group 2}

\item{sd2}{standard deviation of group 2}

\item{sd3}{standard deviation of pair differences}
}
\value{
Deviation Test
}
\description{
\code{meansTestPPair} statistical test for mean values
}
\examples{
# meansTestPPair(25, 50, 10, 60, 15, 10)
}
